requirements.txt
onnxruntime~=1.19.2
torch~=2.2.2
fastapi~=0.115.5
pillow~=10.2.0
uvicorn~=0.15.0
facenet-pytorch~=2.5.2
